refactor: Use ElasticsearchConnectorUtil
